refactor(website): rename component files to dash-case
The site mixed two conventions: docs, scripts, and the playground's non-component
modules were dash-case (building-functions.md, sync-vendor.mjs, _engine.ts,
_walrus.ts, worker.ts) while every .tsx component was PascalCase — most visibly
inside playground/, where _GraphView.tsx sat next to _engine.ts.

Renamed all 19 component files to dash-case, preserving the leading underscore
that marks a file as non-routable in Void's file-based routing:

  _components/  ApiMap → api-map, BuilderHighlight → builder-highlight,
                Button → button, Chip → chip, CodeBlock → code-block,
                CtaBand → cta-band, Footer → footer, Hero → hero,
                LiveHandleStory → live-handle-story, ModuleGraph → module-graph,
                SectionHeader → section-header, Verbs → verbs,
                _CopyButton → _copy-button, _InstallSwitcher → _install-switcher,
                _Reveal → _reveal
  playground/   _DetailPanel → _detail-panel, _GraphView → _graph-view,
                _Playground → _playground, _TreeView → _tree-view

Import specifiers updated to match. The exported component IDENTIFIERS stay
PascalCase (`export default function Hero`, `<CodeBlock />`) — that is the React
convention for components and is not a file name.

Case-only renames (Hero → hero, _Playground → _playground, …) were done as
two-step git mv so they register as real renames on a case-insensitive
filesystem rather than silently no-op'ing.

No route paths change: everything renamed is either under the _components/
directory or carries a leading underscore, so none of it was ever a route.

Verified: tsc --noEmit clean; `npm run build` clean; 20/20 Playwright e2e green;
no remaining PascalCase import specifiers.
