+++ website/pages/_components/_install-switcher.tsx
@@ -1,5 +1,5 @@
 import { useEffect, useState } from 'react'
-import CopyButton from './_CopyButton'
+import CopyButton from './_copy-button'
 
 const cx = (...c: (string | false | undefined)[]) => c.filter(Boolean).join(' ')
