+++ website/pages/_components/code-block.tsx
@@ -1,4 +1,4 @@
-import CopyButton from './_CopyButton'
+import CopyButton from './_copy-button'
 
 const cx = (...c: (string | false | undefined)[]) => c.filter(Boolean).join(' ')
