+++ website/pages/_components/verbs.tsx
@@ -1,6 +1,6 @@
-import CodeBlock from './CodeBlock'
-import SectionHeader from './SectionHeader'
-import Reveal from './_Reveal'
+import CodeBlock from './code-block'
+import SectionHeader from './section-header'
+import Reveal from './_reveal'
 import { inspectSample, editSample, buildSample } from '../_data/samples'
 
 const VERBS = [
