+++ website/pages/playground/_detail-panel.tsx
