+++ website/pages/playground/_playground.tsx
@@ -6,10 +6,10 @@ import type { RunResult } from './_engine'
 import type { InspectResult, GraphNode, Edit, InspectFormat, BuildPresetId, BuildInstrDesc } from './protocol'
 import { BUILD_PRESETS } from './protocol'
 import { WAT_SAMPLES, DEFAULT_WAT } from './_samples'
-import GraphView from './_GraphView'
-import TreeView from './_TreeView'
-import DetailPanel from './_DetailPanel'
-import CodeBlock from '../_components/CodeBlock'
+import GraphView from './_graph-view'
+import TreeView from './_tree-view'
+import DetailPanel from './_detail-panel'
+import CodeBlock from '../_components/code-block'
 
 type Status = 'empty' | 'running' | 'done' | 'error'
 type Mode = 'inspect' | 'edit' | 'build'
