+++ website/pages/playground/index.island.tsx
@@ -1,4 +1,4 @@
-import Playground from './_Playground' with { island: 'load' }
+import Playground from './_playground' with { island: 'load' }
 
 export default function PlaygroundPage() {
   return (
